9th-sprint

+++ test/managers/InMemoryTaskManagerTest.java
@@ -12,6 +12,7 @@ import java.time.Duration;
 import java.time.LocalDateTime;
 import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -26,11 +27,8 @@ class InMemoryTaskManagerTest extends TaskManagerTest<TaskManager>{
         Task task = new Task("Задача", "Описание задачи", Status.NEW,
                 LocalDateTime.of(2023, 9, 2, 7, 0),
                 Duration.ofMinutes(30).toMinutes());
-        int taskId = taskManager.addTask(task);
-        final Task savedTask = taskManager.getTaskById(taskId);
-
-        assertNotNull(savedTask, "Задача не найдена.");
-        assertEquals(task, savedTask, "Задачи не совпадают.");
+        Optional<Task> optionalTask = taskManager.addTask(task);
+        optionalTask.ifPresent(value -> assertEquals(value, task));
 
         final List<Task> tasks = taskManager.getTasksList();
 
@@ -44,9 +42,10 @@ class InMemoryTaskManagerTest extends TaskManagerTest<TaskManager>{
         Task task = new Task("Задача", "Описание задачи", Status.NEW,
                 LocalDateTime.of(2023, 9, 2, 7, 0),
                 Duration.ofMinutes(30).toMinutes());
-        int taskId = taskManager.addTask(task);
+        Optional<Task> optionalTask = taskManager.addTask(task);
         task.setStatus(Status.IN_PROGRESS);
-        assertEquals(Status.IN_PROGRESS, taskManager.getTaskById(taskId).getStatus());
+        optionalTask.ifPresent(value -> assertEquals(Status.IN_PROGRESS, value.getStatus()));
+
     }
 
     @Test
@@ -54,76 +53,83 @@ class InMemoryTaskManagerTest extends TaskManagerTest<TaskManager>{
         Task task = new Task("Задача", "Описание задачи", Status.NEW,
                 LocalDateTime.of(2023, 9, 2, 7, 0),
                 Duration.ofMinutes(30).toMinutes());
-        int taskId = taskManager.addTask(task);
-        assertEquals(task, taskManager.getTaskById(taskId));
-        taskManager.deleteTask(taskId);
+        Optional<Task> optionalTask = taskManager.addTask(task);
+        Task savedTask = taskManager.getTaskById(task.getId()).get();
+        optionalTask.ifPresent(value -> assertEquals(value, task));
+        taskManager.deleteTask(savedTask.getId());
         assertEquals(Collections.EMPTY_LIST, taskManager.getTasksList());
     }
 
     @Test
     void addSubtask() {
         Epic epic = new Epic("Эпик", "Описание эпика");
-        int epicId = taskManager.addEpic(epic);
+        taskManager.addEpic(epic);
         Subtask subtask = new Subtask("Подзадача",
-                "Описание подзадачи", Status.NEW, epicId,
+                "Описание подзадачи", Status.NEW, epic.getId(),
                 LocalDateTime.of(2023, 10, 6, 12, 0),
                 Duration.ofMinutes(30).toMinutes());
-        int subtaskId = taskManager.addSubtask(subtask);
-        final Subtask savedSubtask = taskManager.getSubtaskById(subtaskId);
-        assertNotNull(savedSubtask, "Сабтаски не найдены.");
-        assertEquals(subtask, savedSubtask, "Сабтаски не совпадают.");
-
+        Optional<Subtask> optionalSubtask = taskManager.addSubtask(subtask);
+        optionalSubtask.ifPresent(value -> assertEquals(value, subtask));
         final List<Subtask> subtasks = taskManager.getSubtasksList();
 
         assertNotNull(subtasks, "Сабтаски не возвращаются.");
         assertEquals(1, subtasks.size(), "Неверное количество сабтасков.");
-        assertEquals(savedSubtask, subtasks.get(0), "Сабтаски не совпадают.");
     }
 
     @Test
     void updateSubtask() {
         Epic epic = new Epic("Эпик", "Описание эпика");
-        int epicId = taskManager.addEpic(epic);
+        taskManager.addEpic(epic);
         Subtask subtask = new Subtask("Подзадача",
-                "Описание подзадачи", Status.NEW, epicId,
+                "Описание подзадачи", Status.NEW, epic.getId(),
                 LocalDateTime.of(2023, 10, 6, 12, 0),
                 Duration.ofMinutes(30).toMinutes());
-        int subtaskId = taskManager.addSubtask(subtask);
+        Optional<Subtask> optionalSubtask = taskManager.addSubtask(subtask);
         subtask.setStatus(Status.IN_PROGRESS);
         taskManager.updateSubtask(subtask);
-        assertEquals(Status.IN_PROGRESS, taskManager.getSubtaskById(subtaskId).getStatus());
-        assertEquals(Status.IN_PROGRESS, taskManager.getEpicById(epicId).getStatus());
+        if (optionalSubtask.isPresent()) {
+            Subtask savedSubtask = optionalSubtask.get();
+            assertEquals(Status.IN_PROGRESS, savedSubtask.getStatus());
+            assertEquals(Status.IN_PROGRESS, savedSubtask.getStatus());
+        }
+
     }
 
     @Test
     void deleteEpic() {
         Epic epic = new Epic("Эпик", "Описание эпика");
-        int epicId = taskManager.addEpic(epic);
-        assertEquals(epic, taskManager.getEpicById(epicId));
-        taskManager.deleteEpic(epicId);
+        taskManager.addEpic(epic);
+        Epic savedEpic = taskManager.getEpicById(epic.getId()).get();
+        assertEquals(epic, savedEpic);
+        taskManager.deleteEpic(epic.getId());
         assertEquals(Collections.EMPTY_LIST, taskManager.getEpicsList());
     }
 
     @Test
     void deleteSubtask() {
         Epic epic = new Epic("Эпик", "Описание эпика");
-        int epicId = taskManager.addEpic(epic);
+        Optional<Epic> optionalEpic = taskManager.addEpic(epic);
+        Epic savedEpic = optionalEpic.get();
         Subtask subtask = new Subtask("Подзадача",
-                "Описание подзадачи", Status.NEW, epicId,
+                "Описание подзадачи", Status.NEW, epic.getId(),
                 LocalDateTime.of(2023, 10, 6, 12, 0),
                 Duration.ofMinutes(30).toMinutes());
-        int subtaskId = taskManager.addSubtask(subtask);
-        assertEquals(subtask, taskManager.getSubtaskById(subtaskId));
-        assertEquals(epic, taskManager.getEpicById(epicId));
-        taskManager.deleteSubtask(subtaskId);
-        assertEquals(Collections.EMPTY_LIST, taskManager.getSubtasksList());
+        Optional<Subtask> optionalSubtask = taskManager.addSubtask(subtask);
+        if (optionalSubtask.isPresent()) {
+            Subtask savedSubtask = taskManager.getSubtaskById(subtask.getId()).get();
+            assertEquals(subtask, savedSubtask);
+            assertEquals(epic, savedEpic);
+            taskManager.deleteSubtask(savedSubtask.getId());
+            assertEquals(Collections.EMPTY_LIST, taskManager.getSubtasksList());
+        }
+
     }
 
     @Test
     void addEpic() {
         Epic epic = new Epic("Эпик", "Описание эпика");
-        int epicId = taskManager.addEpic(epic);
-        final Epic savedEpic = taskManager.getEpicById(epicId);
+        Optional<Epic> optionalEpic = taskManager.addEpic(epic);
+        final Epic savedEpic = taskManager.getEpicById(epic.getId()).get();
 
         assertNotNull(savedEpic, "Эпики не найдены.");
         assertEquals(epic, savedEpic, "Эпики не совпадают.");
@@ -140,11 +146,12 @@ class InMemoryTaskManagerTest extends TaskManagerTest<TaskManager>{
         Task task = new Task("Задача", "Описание задачи", Status.NEW,
                 LocalDateTime.of(2023, 9, 2, 7, 0),
                 Duration.ofMinutes(30).toMinutes());
-        int taskId = taskManager.addTask(task);
-        assertEquals(task.getName(), taskManager.getTaskById(taskId).getName());
+        Optional<Task> optionalTask = taskManager.addTask(task);
+        Task savedTask = optionalTask.get();
+        assertEquals(task.getName(), savedTask.getName());
         task.setName("Новое имя задачи");
         task.setDescription("Новое описание задачи");
-        assertEquals(task.getName(), taskManager.getTaskById(taskId).getName());
+        assertEquals(task.getName(),savedTask.getName());
     }
 
     @Test
@@ -152,11 +159,18 @@ class InMemoryTaskManagerTest extends TaskManagerTest<TaskManager>{
         Task task = new Task("Задача", "Описание задачи", Status.NEW,
                 LocalDateTime.of(2023, 9, 2, 7, 0),
                 Duration.ofMinutes(30).toMinutes());
-        int taskId = taskManager.addTask(task);
-        assertEquals(task, taskManager.getTaskById(taskId));
-        task.setName("Updated name");
-        task.setDescription("Updated decp");
-        assertEquals(task, taskManager.getTaskById(taskId));
+        taskManager.addTask(task);
+        Optional<Task> optionalTask = taskManager.getTaskById(task.getId());
+        if(optionalTask.isPresent()) {
+            Task savedTask = optionalTask.get();
+            assertEquals(task, savedTask);
+            task.setName("Updated name");
+            task.setDescription("Updated decp");
+            assertEquals(task, savedTask);
+            task.setName("Updated name");
+            task.setDescription("Updated decp");
+            assertEquals(task, savedTask);
+        }
     }
 
     @Test
@@ -164,10 +178,13 @@ class InMemoryTaskManagerTest extends TaskManagerTest<TaskManager>{
         Task task = new Task("Задача", "Описание задачи", Status.NEW,
                 LocalDateTime.of(2023, 9, 2, 7, 0),
                 Duration.ofMinutes(30).toMinutes());
-        int taskId = taskManager.addTask(task);
-        assertEquals(task.getName(), taskManager.getTaskById(taskId).getName());
-        assertEquals(task.getDescription(), taskManager.getTaskById(taskId).getDescription());
-        assertEquals(task.getStatus(), taskManager.getTaskById(taskId).getStatus());
+        Optional<Task> optionalTask = taskManager.addTask(task);;
+        if(optionalTask.isPresent()) {
+            Task savedTask = optionalTask.get();
+            assertEquals(task.getName(), savedTask.getName());
+            assertEquals(task.getDescription(), savedTask.getDescription());
+            assertEquals(task.getStatus(), savedTask.getStatus());
+        }
     }
 
     @Test
@@ -175,43 +192,57 @@ class InMemoryTaskManagerTest extends TaskManagerTest<TaskManager>{
         Task task = new Task("Задача", "Описание задачи", Status.NEW,
                 LocalDateTime.of(2023, 9, 2, 7, 0),
                 Duration.ofMinutes(30).toMinutes());
-        int taskId = taskManager.addTask(task);
-        assertEquals(task, taskManager.getTaskById(taskId));
-        List<Task> history = taskManager.getHistory();
-        assertArrayEquals(history.toArray(), taskManager.getTasksList().toArray());
-        assertEquals(history.size(), taskManager.getTasksList().size());
-        taskManager.deleteAllTasks();
-        history = taskManager.getTasksList();
-        assertEquals(history.size(), taskManager.getTasksList().size());
+        taskManager.addTask(task);
+        Optional<Task> optionalTask = taskManager.getTaskById(task.getId());
+        if (optionalTask.isPresent()) {
+            Task savedTask = optionalTask.get();
+            assertEquals(task, savedTask);
+            List<Task> history = taskManager.getHistory();
+            assertArrayEquals(history.toArray(), taskManager.getTasksList().toArray());
+            assertEquals(history.size(), taskManager.getTasksList().size());
+            taskManager.deleteAllTasks();
+            history = taskManager.getTasksList();
+            assertEquals(history.size(), taskManager.getTasksList().size());
+        }
+
     }
 
     @Test
     public void deleteEpicFromHistory() {
         Epic epic = new Epic("Эпик", "Описание эпика");
-        int epicId = taskManager.addEpic(epic);
-        assertEquals(epic, taskManager.getEpicById(epicId));
-        List<Task> history = taskManager.getHistory();
-        assertArrayEquals(history.toArray(), taskManager.getEpicsList().toArray());
-        assertEquals(history.size(), taskManager.getEpicsList().size());
-        taskManager.deleteAllEpics();
-        assertEquals(0, taskManager.getEpicsList().size());
+        taskManager.addEpic(epic);
+        Optional<Epic> optionalEpic = taskManager.getEpicById(epic.getId());
+        if (optionalEpic.isPresent()) {
+            Epic savedEpic = optionalEpic.get();
+                    assertEquals(epic, savedEpic);
+            List<Task> history = taskManager.getHistory();
+            assertArrayEquals(history.toArray(), taskManager.getEpicsList().toArray());
+            assertEquals(history.size(), taskManager.getEpicsList().size());
+            taskManager.deleteAllEpics();
+            assertEquals(0, taskManager.getEpicsList().size());
+        }
     }
 
     @Test
     public void deleteSubtasksFromHistory() {
         Epic epic = new Epic("Эпик", "Описание эпика");
-        int epicId = taskManager.addEpic(epic);
+        Optional<Epic> optionalEpic = taskManager.addEpic(epic);
         Subtask subtask = new Subtask("Подзадача",
-                "Описание подзадачи", Status.NEW, epicId,
+                "Описание подзадачи", Status.NEW, epic.getId(),
                 LocalDateTime.of(2023, 10, 6, 12, 0),
                 Duration.ofMinutes(30).toMinutes());
-        int subtaskId = taskManager.addSubtask(subtask);
-        assertEquals(subtask, taskManager.getSubtaskById(subtaskId));
-        List<Task> history = taskManager.getHistory();
-        assertArrayEquals(history.toArray(), taskManager.getSubtasksList().toArray());
-        assertEquals(history.size(), taskManager.getSubtasksList().size());
-        taskManager.deleteAllSubtasks();
-        history = taskManager.getHistory();
-        assertEquals(history.size(), taskManager.getSubtasksList().size());
+        taskManager.addSubtask(subtask);
+        Optional<Subtask> optionalSubtask = taskManager.getSubtaskById(subtask.getId());
+        if (optionalSubtask.isPresent()) {
+            Subtask savedSubtask = optionalSubtask.get();
+            assertEquals(subtask, savedSubtask);
+            List<Task> history = taskManager.getHistory();
+            assertArrayEquals(history.toArray(), taskManager.getSubtasksList().toArray());
+            assertEquals(history.size(), taskManager.getSubtasksList().size());
+            taskManager.deleteAllSubtasks();
+            history = taskManager.getHistory();
+            assertEquals(history.size(), taskManager.getSubtasksList().size());
+        }
+
     }
 }
